Started adding a state to pass into/out of threads
All simulations start from an initial (e.g., random) state in their own
thread. Each thread returns its final state after reaching a defined
terminate condition (for now, `death` or `all food eaten`). From these
final states, the best state(s) is(/are) chosen as base for the genetic
evolution - recombination and/or mutation, and passed into the next
round as new initial states.
For this to work, the respective constructors have to be adjusted to
receive the state, and the Simulator.run() method to return the final
state. Each of the Actors has to update the state itself during runtime,
i.e. needs access to the Simulator threads' _state variable (by-ref).